# cmd/payment/docs/swagger.yaml
basePath: /
definitions:
...
  model.Customer:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      hash_credit_card:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.Merchant:
    properties:
      amount_account:
        type: number
      authentication_key:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para el manejo de pagos
  title: API de Pagos
  version: "1"
paths:
  /api/customer:
    post:
      consumes:
      - application/json
      description: Agrega un nuevo customer a la base de datos con la información
        proporcionada en el cuerpo de la solicitud.
      parameters:
      - description: Información del customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/model.Customer'
      produces:
      - application/json
      responses:
        "201":
          description: Customer creado exitosamente
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Error al procesar los datos de entrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor al crear el customer
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Crea un nuevo customer
      tags:
      - customer
  /api/customer/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un customer por su ID
      parameters:
      - description: ID del Customer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Customer eliminado exitosamente
        "404":
          description: Customer no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Elimina un customer
      tags:
      - customer
    get:
      consumes:
      - application/json
      description: Obtiene un customer por su ID
      parameters:
      - description: ID del Customer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Customer encontrado
          schema:
            $ref: '#/definitions/model.Customer'
        "404":
          description: Customer no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene un customer
      tags:
      - customer
    put:
      consumes:
      - application/json
      description: Actualiza la información de un customer existente por su ID
      parameters:
      - description: ID del Customer
        in: path
        name: id
        required: true
        type: integer
      - description: Información actualizada del Customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/model.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: Customer actualizado
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Entrada inválida
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Customer no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Actualiza un customer
      tags:
      - customer
  /api/merchant:
    post:
      consumes:
      - application/json
      description: Agrega un nuevo merchant a la base de datos
      parameters:
      - description: Información del merchant
        in: body
        name: merchant
        required: true
        schema:
          $ref: '#/definitions/model.Merchant'
      produces:
      - application/json
      responses:
        "201":
          description: Merchant creado
          schema:
            $ref: '#/definitions/model.Merchant'
        "400":
          description: Entrada inválida
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Crea un nuevo merchant
      tags:
      - merchant
  /api/merchant/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un merchant por su ID
      parameters:
      - description: ID del Merchant
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Merchant eliminado exitosamente
        "404":
          description: Merchant no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Elimina un merchant
      tags:
      - merchant
    get:
      consumes:
      - application/json
      description: Obtiene un merchant por su ID
      parameters:
      - description: ID del Merchant
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Merchant'
        "404":
          description: Merchant no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene un merchant
      tags:
      - merchant
    put:
      consumes:
      - application/json
      description: Actualiza la información de un merchant existente por su ID
      parameters:
      - description: ID del Merchant
        in: path
        name: id
        required: true
        type: integer
      - description: Información actualizada del Merchant
        in: body
        name: merchant
        required: true
        schema:
          $ref: '#/definitions/model.Merchant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Merchant'
        "400":
          description: Entrada inválida
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Merchant no encontrado
          schema:
            additionalProperties: true
            type: object
      summary: Actualiza un merchant
      tags:
      - merchant
...
swagger: "2.0"